Request a RetroGuard release v3.6.6
Because Retroguard packages the Signature ASM classes, ASM has been crashing with a MethodNotFOundException. In recent commits to Retroguard, these packaged classes have been shadowed in a way where they will no longer cause a problem with SpecialSource.

For that reason, I reuquest that this PR is pulled in order to signal to jenkins that this should be uploaded as a new version rather than a snapshot. Then the vesion will be accessible to me in ForgeGradle.